refactor(types): 将 API 定义移至新文件 api-endpoints.ts
- 在 packages/types/src 目录下新建 api-endpoints.ts 文件
- 将原有的 API 定义从 common.ts 移到 api-endpoints.ts
- 更新 package.json 和 tsup.config.ts 以包含新文件
- 从 common.ts 中移除已迁移的 API 定义
